BIB-522: changed to make dependency inj easier
